Move SuggestionList processing methods inside class
